requirements.txt
git+https://github.com/Rapptz/discord.py@async
PyYAML==3.11
redis==2.10.5
youtube-dl==2016.4.13
soundcloud==0.5.0
